+++ src/Lykke.Service.OperationsRepository.AzureRepositories/CashOperations/ClientTradesRepository.cs
@@ -48,7 +48,7 @@ namespace Lykke.Service.OperationsRepository.AzureRepositories.CashOperations
         }
         public double Volume { get; set; }
         public string ClientId { get; set; }
-        public double FeeSize { get; set; }
+        public decimal FeeSize { get; set; }
         public FeeType FeeType { get; set; }
         public string FeeTypeText
         {
